requirements.txt
# PDF Processing
pdfplumber>=0.11.0
PyMuPDF>=1.24.0

# OCR (Optical Character Recognition)
pytesseract>=0.3.10
Pillow>=10.0.0

# GUI Framework
customtkinter>=5.2.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Code Quality
ruff>=0.1.0

# Packaging
pyinstaller>=6.0.0
